Forgot commit message on last: added AMQPTransport and osis queue worker. Now there is a queueing system between appserver-WFE and WFE-osis.
Some todo's: every appserver/wfe instance has to get a guid for the return queue.
We should change the DRPClient in the DRP: this now only allows to send 1 message and waits for the result. It's more logical to push messages on the osis queues and let the osis_worker handle the hard work.